Move method to change display order to concerns/display_order
